[deba] make order gossip work with new changes
