+++ chain.json
@@ -5,12 +5,20 @@
   "tx-regossip-max-size": 32,
   "priority-regossip-max-txs": 500,
   "priority-regossip-txs-per-address": 200,
-  "priority-regossip-addresses": ["0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266", "0x70997970C51812dc3A010C7d01b50e0d17dc79C8", "0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC", "0x8db97C7cEcE249c2b98bDC0226Cc4C2A57BF52FC"],
+  "priority-regossip-addresses": [
+    "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+    "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+    "0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC",
+    "0x8db97C7cEcE249c2b98bDC0226Cc4C2A57BF52FC"
+  ],
   "validator-private-key-file": "/tmp/validator.pk",
   "is-validator": true,
   "trading-api-enabled": true,
   "testing-api-enabled": true,
   "load-from-snapshot-enabled": true,
   "snapshot-file-path": "/tmp/snapshot",
-  "makerbook-database-path": "/tmp/makerbook"
+  "makerbook-database-path": "/tmp/makerbook",
+  "order-gossip-num-validators": 10,
+  "order-gossip-num-non-validators": 5,
+  "order-gossip-num-peers": 15
 }
