+++ network-configs/aylin/chain.json
@@ -5,10 +5,15 @@
   "tx-regossip-max-size": 32,
   "priority-regossip-max-txs": 32,
   "priority-regossip-txs-per-address": 20,
-  "priority-regossip-addresses": ["0x06CCAD927e6B1d36E219Cb582Af3185D0705f78F"],
+  "priority-regossip-addresses": [
+    "0x06CCAD927e6B1d36E219Cb582Af3185D0705f78F"
+  ],
   "validator-private-key-file": "/home/ubuntu/validator.pk",
   "feeRecipient": "<insert address corresponding to private key above>",
   "is-validator": true,
   "snapshot-file-path": "/tmp/snapshot",
-  "makerbook-database-path": "/tmp/makerbook"
+  "makerbook-database-path": "/tmp/makerbook",
+  "order-gossip-num-validators": 10,
+  "order-gossip-num-non-validators": 5,
+  "order-gossip-num-peers": 15
 }
