+++ network-configs/aylin/t2WSjSsoE3geV9ARu5r7gzTc5UayePy3NxDrSTx7hadLYvqbg.json
@@ -1,6 +1,3 @@
 {
-    "proposerMinBlockDelay": 0,
-    "appGossipValidatorSize": 10,
-    "appGossipNonValidatorSize": 5,
-    "appGossipPeerSize": 15
+    "proposerMinBlockDelay": 0
 }
