+++ network-configs/hubblenet/2mxZY7A2t1tuRMALW4BcBUPVGNR3LH1DXhftdbLAHm1QtDkFp8.json
@@ -1,6 +1,3 @@
 {
-    "proposerMinBlockDelay": 0,
-    "appGossipValidatorSize": 10,
-    "appGossipNonValidatorSize": 5,
-    "appGossipPeerSize": 15
+    "proposerMinBlockDelay": 0
 }
