+++ plugin/evm/gossip_stats.go
@@ -10,29 +10,15 @@ var _ GossipStats = &gossipStats{}
 // GossipStats contains methods for updating incoming and outgoing gossip stats.
 type GossipStats interface {
 	IncEthTxsGossipReceived()
-
-	// new vs. known txs received
 	IncEthTxsGossipReceivedError()
 	IncEthTxsGossipReceivedKnown()
 	IncEthTxsGossipReceivedNew()
 
 	IncSignedOrdersGossipReceived(count int64)
 	IncSignedOrdersGossipBatchReceived()
-
-	// new vs. known txs received
 	IncSignedOrdersGossipReceivedKnown()
 	IncSignedOrdersGossipReceivedNew()
 	IncSignedOrdersGossipReceiveError()
-}
-
-// GossipSentStats groups functions for outgoing gossip stats.
-type GossipSentStats interface {
-	IncEthTxsGossipSent()
-
-	// regossip
-	IncEthTxsRegossipQueued()
-	IncEthTxsRegossipQueuedLocal(count int)
-	IncEthTxsRegossipQueuedRemote(count int)
 
 	IncSignedOrdersGossipSent(count int64)
 	IncSignedOrdersGossipBatchSent()
@@ -42,27 +28,21 @@ type GossipSentStats interface {
 
 // gossipStats implements stats for incoming and outgoing gossip stats.
 type gossipStats struct {
-	// messages
-	ethTxsGossipReceived metrics.Counter
-
-	// new vs. known txs received
+	ethTxsGossipReceived      metrics.Counter
 	ethTxsGossipReceivedError metrics.Counter
 	ethTxsGossipReceivedKnown metrics.Counter
 	ethTxsGossipReceivedNew   metrics.Counter
 
-	// messages
-	signedOrdersGossipSent          metrics.Counter
-	signedOrdersGossipBatchSent     metrics.Counter
-	signedOrdersGossipSendError     metrics.Counter
-	signedOrdersGossipOrderExpired  metrics.Counter
 	signedOrdersGossipReceived      metrics.Counter
 	signedOrdersGossipBatchReceived metrics.Counter
-
-	// regossip
-	// new vs. known txs received
 	signedOrdersGossipReceivedKnown metrics.Counter
 	signedOrdersGossipReceivedNew   metrics.Counter
 	signedOrdersGossipReceiveError  metrics.Counter
+
+	signedOrdersGossipSent         metrics.Counter
+	signedOrdersGossipBatchSent    metrics.Counter
+	signedOrdersGossipSendError    metrics.Counter
+	signedOrdersGossipOrderExpired metrics.Counter
 }
 
 func NewGossipStats() GossipStats {
