+++ scripts/upgrade_local.sh
@@ -9,6 +9,6 @@ avalanche network stop --snapshot-name snap1
 avalanche subnet upgrade vm localnet --binary custom_evm.bin --local
 
 # utse tee to keep showing outut while storing in a var
-OUTPUT=$(avalanche network start --avalanchego-version v1.11.1 --snapshot-name snap1 --config .avalanche-cli.json | tee /dev/fd/2)
+OUTPUT=$(avalanche network start --avalanchego-version v1.11.2 --snapshot-name snap1 --config .avalanche-cli.json | tee /dev/fd/2)
 
 setStatus
